feat: loop for array of list of objects
